Add Repositories, Dependency Injection and  DTO
