+++ ClinicApp/Controllers/DoctorController.cs
@@ -1,25 +1,29 @@
+using Repositories.Doctor.Base;
+using Repositories.Patient.Base;
+
 namespace ClinicApp.Controllers;
 
-
-using ClinicApp.ClinicDB;
 using Microsoft.AspNetCore.Mvc;
-
 
 public class DoctorController : Controller
 {
+    private readonly IDoctorRepository doctorRepository;
+    public DoctorController(IDoctorRepository doctorRepository)
+    {
+        this.doctorRepository = doctorRepository;
+    }
+
     [HttpGet]
     public async Task<IActionResult> ShowDoctorByMedicalDepartment(int medicalDepartment)
     {
-        var DB = new BonadeaDB();
-        var doctors = await DB.GetDoctorsByMedicalDepartment(medicalDepartment);
+        var doctors = await doctorRepository.GetDoctorsByMedicalDepartment(medicalDepartment);
         return View(model: doctors);
     }
 
     [HttpGet]
     public async Task<IActionResult> ShowAllDoctors()
     {
-        var DB = new BonadeaDB();
-        var doctors = await DB.GetAllDoctors();
+        var doctors = await doctorRepository.GetAllDoctors();
         return View(model: doctors);
     
     }
