+++ ClinicApp/Controllers/PatientController.cs
@@ -1,6 +1,8 @@
+using Repositories.Patient.Base;
+using Repositories.Doctor.Base;
+
 namespace ClinicApp.Controllers;
 
-using ClinicApp.ClinicDB;
 using ClinicApp.Models;
 using ClinicApp.ViewModels;
 using Microsoft.AspNetCore.Mvc;
@@ -8,6 +10,14 @@ using Newtonsoft.Json;
 
 public class PatientController : Controller
 {
+    private readonly IPatientRepository patientRepository;
+    private readonly IDoctorRepository doctorRepository;
+    public PatientController(IPatientRepository patientRepository, IDoctorRepository doctorRepository)
+    {
+        this.patientRepository = patientRepository;
+        this.doctorRepository = doctorRepository;
+    }
+
     [HttpPost]
     public IActionResult CreatePatient([FromForm(Name = "doctor")] string doctorJson)
     {
@@ -19,21 +29,18 @@ public class PatientController : Controller
     [HttpGet]
     public async Task<IActionResult> ShowAllPatients()
     {
-        var DB = new BonadeaDB();
-        var patients = await DB.GetAllPatients();
+        var patients = await patientRepository.GetAllPatients();
         return View(model: patients);
     }
 
     [HttpGet]
     public async Task<IActionResult> ShowPatientsByDoctor([FromQuery] string doctorFIN)
     {
-        var DB = new BonadeaDB();
-        var patients = await DB.GetPatientsByDoctor(doctorFIN);
-        var doctor = await DB.GetDoctorByFIN(doctorFIN);
+        var patients = await patientRepository.GetPatientsByDoctor(doctorFIN);
+        var doctor = await doctorRepository.GetDoctorByFIN(doctorFIN);
         var viewModelPatientsByDoctor = new ViewModelPatientsByDoctor();
         viewModelPatientsByDoctor.doctor = doctor;
         viewModelPatientsByDoctor.patients = patients;
-
         return View(model: viewModelPatientsByDoctor);
     }
 }
