+++ ClinicApp/DTO/PatientDTO.cs
@@ -1,12 +1,9 @@
 namespace ClinicApp.DTO;
 
-using ClinicApp.Models;
-
-
-public class DoctorDTO
+public class PatientDTO
 {
     public string? FIN { get; set; }
     public string? Firstname { get; set; }
     public string? Lastname { get; set; }
-    public MedicalDepartmentsEnum? MedicalDepartment { get; set; }
+    public string? Email { get; set; }
 }
