+++ ClinicApp/Program.cs
@@ -1,25 +1,35 @@
+using ClinicApp.Repositories.Doctor;
+using ClinicApp.Repositories.Patient;
+using Repositories.Doctor.Base;
+using Repositories.Patient.Base;
+
 var builder = WebApplication.CreateBuilder(args);
 
-// Add services to the container.
+
 builder.Services.AddControllersWithViews();
+
+
+
+builder.Services.AddSingleton<IDoctorRepository, DoctorRepository>();
+builder.Services.AddSingleton<IPatientRepository, PatientRepository>();
 
 var app = builder.Build();
 
-// Configure the HTTP request pipeline.
+
+
+
+
+
 if (!app.Environment.IsDevelopment())
 {
     app.UseExceptionHandler("/Home/Error");
-    // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
     app.UseHsts();
 }
 
 app.UseHttpsRedirection();
 app.UseStaticFiles();
-
 app.UseRouting();
-
 app.UseAuthorization();
-
 app.MapDefaultControllerRoute();
 
 app.Run();
