+++ ClinicApp/Repositories/Doctor/Base/IDoctorRepository.cs
@@ -1,0 +1,11 @@
+namespace Repositories.Doctor.Base;
+
+using ClinicApp.Models;
+
+public interface IDoctorRepository
+{
+    Task<IEnumerable<Doctor>> GetDoctorsByMedicalDepartment(int medicalDepartment);
+    Task<Doctor> GetDoctorByFIN(string? doctorFIN);
+    Task AddPatientToDoctor(int doctorId, int patientId);
+    Task<IEnumerable<Doctor>> GetAllDoctors();
+}
